@@ -1,3 +1,5 @@
+'use strict';
+
 const EventEmitter = require('events');
 
 /**
@@ -5,43 +7,52 @@ const EventEmitter = require('events');
  * @extends {EventEmitter}
  */
 class VolumeInterface extends EventEmitter {
-  constructor({ volume = 0 } = {}) {
+  constructor({ volume = 1 } = {}) {
     super();
-    this.setVolume(volume || 1);
+    this.setVolume(volume);
   }
 
   /**
-   * The current volume of the broadcast
+   * Whether or not the volume of this stream is editable
+   * @type {boolean}
    * @readonly
+   */
+  get volumeEditable() {
+    return true;
+  }
+
+  /**
+   * The current volume of the stream
    * @type {number}
+   * @readonly
    */
   get volume() {
     return this._volume;
   }
 
   /**
-   * The current volume of the broadcast in decibels
-   * @readonly
+   * The current volume of the stream in decibels
    * @type {number}
+   * @readonly
    */
   get volumeDecibels() {
-    return Math.log10(this._volume) * 20;
+    return Math.log10(this.volume) * 20;
   }
 
   /**
-   * The current volume of the broadcast from a logarithmic scale
-   * @readonly
+   * The current volume of the stream from a logarithmic scale
    * @type {number}
+   * @readonly
    */
   get volumeLogarithmic() {
-    return Math.pow(this._volume, 1 / 1.660964);
+    return Math.pow(this.volume, 1 / 1.660964);
   }
 
   applyVolume(buffer, volume) {
     volume = volume || this._volume;
     if (volume === 1) return buffer;
 
-    const out = new Buffer(buffer.length);
+    const out = Buffer.alloc(buffer.length);
     for (let i = 0; i < buffer.length; i += 2) {
       if (i >= buffer.length - 1) break;
       const uint = Math.min(32767, Math.max(-32767, Math.floor(volume * buffer.readInt16LE(i))));
@@ -67,7 +78,7 @@ class VolumeInterface extends EventEmitter {
   }
 
   /**
-   * Set the volume in decibels.
+   * Sets the volume in decibels.
    * @param {number} db The decibels
    */
   setVolumeDecibels(db) {
@@ -75,7 +86,7 @@ class VolumeInterface extends EventEmitter {
   }
 
   /**
-   * Set the volume so that a perceived value of 0.5 is half the perceived volume etc.
+   * Sets the volume so that a perceived value of 0.5 is half the perceived volume etc.
    * @param {number} value The value for the volume
    */
   setVolumeLogarithmic(value) {
@@ -83,4 +94,10 @@ class VolumeInterface extends EventEmitter {
   }
 }
 
-module.exports = VolumeInterface;
+const props = ['volumeDecibels', 'volumeLogarithmic', 'setVolumeDecibels', 'setVolumeLogarithmic'];
+
+exports.applyToClass = function applyToClass(structure) {
+  for (const prop of props) {
+    Object.defineProperty(structure.prototype, prop, Object.getOwnPropertyDescriptor(VolumeInterface.prototype, prop));
+  }
+};
