@@ -1,34 +1,33 @@
+'use strict';
+
 const Action = require('./Action');
-const Constants = require('../../util/Constants');
-const Util = require('../../util/Util');
+const Channel = require('../../structures/Channel');
+const { ChannelTypes } = require('../../util/Constants');
 
 class ChannelUpdateAction extends Action {
   handle(data) {
     const client = this.client;
 
-    const channel = client.channels.get(data.id);
+    let channel = client.channels.cache.get(data.id);
     if (channel) {
-      const oldChannel = Util.cloneObject(channel);
-      channel.setup(data);
-      client.emit(Constants.Events.CHANNEL_UPDATE, oldChannel, channel);
+      const old = channel._update(data);
+
+      if (ChannelTypes[channel.type.toUpperCase()] !== data.type) {
+        const newChannel = Channel.create(this.client, data, channel.guild);
+        for (const [id, message] of channel.messages.cache) newChannel.messages.cache.set(id, message);
+        newChannel._typing = new Map(channel._typing);
+        channel = newChannel;
+        this.client.channels.cache.set(channel.id, channel);
+      }
+
       return {
-        old: oldChannel,
+        old,
         updated: channel,
       };
     }
 
-    return {
-      old: null,
-      updated: null,
-    };
+    return {};
   }
 }
-
-/**
- * Emitted whenever a channel is updated - e.g. name change, topic change.
- * @event Client#channelUpdate
- * @param {Channel} oldChannel The channel before the update
- * @param {Channel} newChannel The channel after the update
- */
 
 module.exports = ChannelUpdateAction;
