@@ -1,17 +1,20 @@
+'use strict';
+
 const Action = require('./Action');
+const { Events } = require('../../util/Constants');
 
 class GuildEmojiDeleteAction extends Action {
   handle(emoji) {
-    const client = this.client;
-    client.dataManager.killEmoji(emoji);
+    emoji.guild.emojis.cache.delete(emoji.id);
+    emoji.deleted = true;
+    /**
+     * Emitted whenever a custom emoji is deleted in a guild.
+     * @event Client#emojiDelete
+     * @param {GuildEmoji} emoji The emoji that was deleted
+     */
+    this.client.emit(Events.GUILD_EMOJI_DELETE, emoji);
     return { emoji };
   }
 }
-
-/**
- * Emitted whenever a custom guild emoji is deleted.
- * @event Client#emojiDelete
- * @param {Emoji} emoji The emoji that was deleted
- */
 
 module.exports = GuildEmojiDeleteAction;
