@@ -1,24 +1,20 @@
-const Action = require('./Action');
+'use strict';
 
-function mappify(iterable) {
-  const map = new Map();
-  for (const x of iterable) map.set(...x);
-  return map;
-}
+const Action = require('./Action');
 
 class GuildEmojisUpdateAction extends Action {
   handle(data) {
-    const guild = this.client.guilds.get(data.guild_id);
+    const guild = this.client.guilds.cache.get(data.guild_id);
     if (!guild || !guild.emojis) return;
 
-    const deletions = mappify(guild.emojis.entries());
+    const deletions = new Map(guild.emojis.cache);
 
     for (const emoji of data.emojis) {
       // Determine type of emoji event
-      const cachedEmoji = guild.emojis.get(emoji.id);
+      const cachedEmoji = guild.emojis.cache.get(emoji.id);
       if (cachedEmoji) {
         deletions.delete(emoji.id);
-        if (!cachedEmoji.equals(emoji, true)) {
+        if (!cachedEmoji.equals(emoji)) {
           // Emoji updated
           this.client.actions.GuildEmojiUpdate.handle(cachedEmoji, emoji);
         }
