@@ -1,25 +1,30 @@
+'use strict';
+
 const Action = require('./Action');
-const Constants = require('../../util/Constants');
-const Util = require('../../util/Util');
+const { Events } = require('../../util/Constants');
 
 class GuildRoleUpdateAction extends Action {
   handle(data) {
     const client = this.client;
-    const guild = client.guilds.get(data.guild_id);
+    const guild = client.guilds.cache.get(data.guild_id);
 
     if (guild) {
-      const roleData = data.role;
-      let oldRole = null;
+      let old = null;
 
-      const role = guild.roles.get(roleData.id);
+      const role = guild.roles.cache.get(data.role.id);
       if (role) {
-        oldRole = Util.cloneObject(role);
-        role.setup(data.role);
-        client.emit(Constants.Events.GUILD_ROLE_UPDATE, oldRole, role);
+        old = role._update(data.role);
+        /**
+         * Emitted whenever a guild role is updated.
+         * @event Client#roleUpdate
+         * @param {Role} oldRole The role before the update
+         * @param {Role} newRole The role after the update
+         */
+        client.emit(Events.GUILD_ROLE_UPDATE, old, role);
       }
 
       return {
-        old: oldRole,
+        old,
         updated: role,
       };
     }
@@ -30,12 +35,5 @@ class GuildRoleUpdateAction extends Action {
     };
   }
 }
-
-/**
- * Emitted whenever a guild role is updated.
- * @event Client#roleUpdate
- * @param {Role} oldRole The role before the update
- * @param {Role} newRole The role after the update
- */
 
 module.exports = GuildRoleUpdateAction;
