@@ -1,14 +1,16 @@
+'use strict';
+
 const Action = require('./Action');
 
 class GuildRolesPositionUpdate extends Action {
   handle(data) {
     const client = this.client;
 
-    const guild = client.guilds.get(data.guild_id);
+    const guild = client.guilds.cache.get(data.guild_id);
     if (guild) {
       for (const partialRole of data.roles) {
-        const role = guild.roles.get(partialRole.id);
-        if (role) role.position = partialRole.position;
+        const role = guild.roles.cache.get(partialRole.id);
+        if (role) role.rawPosition = partialRole.position;
       }
     }
 
