@@ -1,54 +1,38 @@
+'use strict';
+
 const Action = require('./Action');
-const Message = require('../../structures/Message');
+const { Events } = require('../../util/Constants');
 
 class MessageCreateAction extends Action {
   handle(data) {
     const client = this.client;
-
-    const channel = client.channels.get((data instanceof Array ? data[0] : data).channel_id);
-    const user = client.users.get((data instanceof Array ? data[0] : data).author.id);
+    const channel = client.channels.cache.get(data.channel_id);
     if (channel) {
-      const member = channel.guild ? channel.guild.member(user) : null;
-      if (data instanceof Array) {
-        const messages = new Array(data.length);
-        for (let i = 0; i < data.length; i++) {
-          messages[i] = channel._cacheMessage(new Message(channel, data[i], client));
-        }
-        const lastMessage = messages[messages.length - 1];
-        channel.lastMessageID = lastMessage.id;
-        channel.lastMessage = lastMessage;
-        if (user) {
-          user.lastMessageID = lastMessage.id;
-          user.lastMessage = lastMessage;
-        }
-        if (member) {
-          member.lastMessageID = lastMessage.id;
-          member.lastMessage = lastMessage;
-        }
-        return {
-          messages,
-        };
-      } else {
-        const message = channel._cacheMessage(new Message(channel, data, client));
-        channel.lastMessageID = data.id;
-        channel.lastMessage = message;
-        if (user) {
-          user.lastMessageID = data.id;
-          user.lastMessage = message;
-        }
-        if (member) {
-          member.lastMessageID = data.id;
-          member.lastMessage = message;
-        }
-        return {
-          message,
-        };
+      const existing = channel.messages.cache.get(data.id);
+      if (existing) return { message: existing };
+      const message = channel.messages.add(data);
+      const user = message.author;
+      let member = message.member;
+      channel.lastMessageID = data.id;
+      if (user) {
+        user.lastMessageID = data.id;
+        user.lastMessageChannelID = channel.id;
       }
+      if (member) {
+        member.lastMessageID = data.id;
+        member.lastMessageChannelID = channel.id;
+      }
+
+      /**
+       * Emitted whenever a message is created.
+       * @event Client#message
+       * @param {Message} message The created message
+       */
+      client.emit(Events.MESSAGE_CREATE, message);
+      return { message };
     }
 
-    return {
-      message: null,
-    };
+    return {};
   }
 }
 
