@@ -1,40 +1,24 @@
+'use strict';
+
 const Action = require('./Action');
-const Constants = require('../../util/Constants');
 
 class MessageUpdateAction extends Action {
   handle(data) {
-    const client = this.client;
-
-    const channel = client.channels.get(data.channel_id);
+    const channel = this.getChannel(data);
     if (channel) {
-      const message = channel.messages.get(data.id);
+      const { id, channel_id, guild_id, author, timestamp, type } = data;
+      const message = this.getMessage({ id, channel_id, guild_id, author, timestamp, type }, channel);
       if (message) {
-        message.patch(data);
-        client.emit(Constants.Events.MESSAGE_UPDATE, message._edits[0], message);
+        const old = message.patch(data);
         return {
-          old: message._edits[0],
+          old,
           updated: message,
         };
       }
-
-      return {
-        old: message,
-        updated: message,
-      };
     }
 
-    return {
-      old: null,
-      updated: null,
-    };
+    return {};
   }
 }
-
-/**
- * Emitted whenever a message is updated - e.g. embed or content change.
- * @event Client#messageUpdate
- * @param {Message} oldMessage The message before the update
- * @param {Message} newMessage The message after the update
- */
 
 module.exports = MessageUpdateAction;
