@@ -1,24 +1,69 @@
+'use strict';
+
+const { Events } = require('../util/Constants');
 const Util = require('../util/Util');
 
 /**
- * Helper class for sharded clients spawned as a child process, such as from a ShardingManager.
+ * Helper class for sharded clients spawned as a child process/worker, such as from a {@link ShardingManager}.
+ * Utilises IPC to send and receive data to/from the master process and other shards.
  */
 class ShardClientUtil {
   /**
-   * @param {Client} client The client of the current shard
+   * @param {Client} client Client of the current shard
+   * @param {ShardingManagerMode} mode Mode the shard was spawned with
    */
-  constructor(client) {
+  constructor(client, mode) {
+    /**
+     * Client for the shard
+     * @type {Client}
+     */
     this.client = client;
-    process.on('message', this._handleMessage.bind(this));
+
+    /**
+     * Mode the shard was spawned with
+     * @type {ShardingManagerMode}
+     */
+    this.mode = mode;
+
+    /**
+     * Message port for the master process (only when {@link ShardClientUtil#mode} is `worker`)
+     * @type {?MessagePort}
+     */
+    this.parentPort = null;
+
+    if (mode === 'process') {
+      process.on('message', this._handleMessage.bind(this));
+      client.on('ready', () => {
+        process.send({ _ready: true });
+      });
+      client.on('disconnect', () => {
+        process.send({ _disconnect: true });
+      });
+      client.on('reconnecting', () => {
+        process.send({ _reconnecting: true });
+      });
+    } else if (mode === 'worker') {
+      this.parentPort = require('worker_threads').parentPort;
+      this.parentPort.on('message', this._handleMessage.bind(this));
+      client.on('ready', () => {
+        this.parentPort.postMessage({ _ready: true });
+      });
+      client.on('disconnect', () => {
+        this.parentPort.postMessage({ _disconnect: true });
+      });
+      client.on('reconnecting', () => {
+        this.parentPort.postMessage({ _reconnecting: true });
+      });
+    }
   }
 
   /**
-   * ID of this shard
-   * @type {number}
+   * Array of shard IDs of this client
+   * @type {number[]}
    * @readonly
    */
-  get id() {
-    return this.client.options.shardId;
+  get ids() {
+    return this.client.options.shards;
   }
 
   /**
@@ -34,62 +79,93 @@ class ShardClientUtil {
    * Sends a message to the master process.
    * @param {*} message Message to send
    * @returns {Promise<void>}
+   * @emits Shard#message
    */
   send(message) {
     return new Promise((resolve, reject) => {
-      const sent = process.send(message, err => {
-        if (err) reject(err); else resolve();
-      });
-      if (!sent) throw new Error('Failed to send message to master process.');
+      if (this.mode === 'process') {
+        process.send(message, err => {
+          if (err) reject(err);
+          else resolve();
+        });
+      } else if (this.mode === 'worker') {
+        this.parentPort.postMessage(message);
+        resolve();
+      }
     });
   }
 
   /**
    * Fetches a client property value of each shard.
    * @param {string} prop Name of the client property to get, using periods for nesting
-   * @returns {Promise<Array>}
+   * @returns {Promise<Array<*>>}
    * @example
-   * client.shard.fetchClientValues('guilds.size')
-   *   .then(results => {
-   *     console.log(`${results.reduce((prev, val) => prev + val, 0)} total guilds`);
-   *   })
+   * client.shard.fetchClientValues('guilds.cache.size')
+   *   .then(results => console.log(`${results.reduce((prev, val) => prev + val, 0)} total guilds`))
    *   .catch(console.error);
+   * @see {@link ShardingManager#fetchClientValues}
    */
   fetchClientValues(prop) {
     return new Promise((resolve, reject) => {
+      const parent = this.parentPort || process;
+
       const listener = message => {
         if (!message || message._sFetchProp !== prop) return;
-        process.removeListener('message', listener);
-        if (!message._error) resolve(message._result); else reject(Util.makeError(message._error));
+        parent.removeListener('message', listener);
+        if (!message._error) resolve(message._result);
+        else reject(Util.makeError(message._error));
       };
-      process.on('message', listener);
+      parent.on('message', listener);
 
       this.send({ _sFetchProp: prop }).catch(err => {
-        process.removeListener('message', listener);
+        parent.removeListener('message', listener);
         reject(err);
       });
     });
   }
 
   /**
-   * Evaluates a script on all shards, in the context of the Clients.
-   * @param {string} script JavaScript to run on each shard
-   * @returns {Promise<Array>} Results of the script execution
+   * Evaluates a script or function on all shards, in the context of the {@link Client}s.
+   * @param {string|Function} script JavaScript to run on each shard
+   * @returns {Promise<Array<*>>} Results of the script execution
+   * @example
+   * client.shard.broadcastEval('this.guilds.cache.size')
+   *   .then(results => console.log(`${results.reduce((prev, val) => prev + val, 0)} total guilds`))
+   *   .catch(console.error);
+   * @see {@link ShardingManager#broadcastEval}
    */
   broadcastEval(script) {
     return new Promise((resolve, reject) => {
+      const parent = this.parentPort || process;
+      script = typeof script === 'function' ? `(${script})(this)` : script;
+
       const listener = message => {
         if (!message || message._sEval !== script) return;
-        process.removeListener('message', listener);
-        if (!message._error) resolve(message._result); else reject(Util.makeError(message._error));
+        parent.removeListener('message', listener);
+        if (!message._error) resolve(message._result);
+        else reject(Util.makeError(message._error));
       };
-      process.on('message', listener);
+      parent.on('message', listener);
 
       this.send({ _sEval: script }).catch(err => {
-        process.removeListener('message', listener);
+        parent.removeListener('message', listener);
         reject(err);
       });
     });
+  }
+
+  /**
+   * Requests a respawn of all shards.
+   * @param {number} [shardDelay=5000] How long to wait between shards (in milliseconds)
+   * @param {number} [respawnDelay=500] How long to wait between killing a shard's process/worker and restarting it
+   * (in milliseconds)
+   * @param {number} [spawnTimeout=30000] The amount in milliseconds to wait for a shard to become ready before
+   * continuing to another. (-1 or Infinity for no wait)
+   * @returns {Promise<void>} Resolves upon the message being sent
+   * @see {@link ShardingManager#respawnAll}
+   */
+  respawnAll(shardDelay = 5000, respawnDelay = 500, spawnTimeout = 30000) {
+    return this.send({ _sRespawnAll: { shardDelay, respawnDelay, spawnTimeout } });
   }
 
   /**
@@ -97,7 +173,7 @@ class ShardClientUtil {
    * @param {*} message Message received
    * @private
    */
-  _handleMessage(message) {
+  async _handleMessage(message) {
     if (!message) return;
     if (message._fetchProp) {
       const props = message._fetchProp.split('.');
@@ -106,7 +182,7 @@ class ShardClientUtil {
       this._respond('fetchProp', { _fetchProp: message._fetchProp, _result: value });
     } else if (message._eval) {
       try {
-        this._respond('eval', { _eval: message._eval, _result: this.client._eval(message._eval) });
+        this._respond('eval', { _eval: message._eval, _result: await this.client._eval(message._eval) });
       } catch (err) {
         this._respond('eval', { _eval: message._eval, _error: Util.makePlainError(err) });
       }
@@ -122,20 +198,29 @@ class ShardClientUtil {
   _respond(type, message) {
     this.send(message).catch(err => {
       err.message = `Error when sending ${type} response to master process: ${err.message}`;
-      this.client.emit('error', err);
+      /**
+       * Emitted when the client encounters an error.
+       * @event Client#error
+       * @param {Error} error The error encountered
+       */
+      this.client.emit(Events.ERROR, err);
     });
   }
 
   /**
    * Creates/gets the singleton of this class.
    * @param {Client} client The client to use
+   * @param {ShardingManagerMode} mode Mode the shard was spawned with
    * @returns {ShardClientUtil}
    */
-  static singleton(client) {
+  static singleton(client, mode) {
     if (!this._singleton) {
-      this._singleton = new this(client);
+      this._singleton = new this(client, mode);
     } else {
-      client.emit('warn', 'Multiple clients created in child process; only the first will handle sharding helpers.');
+      client.emit(
+        Events.WARN,
+        'Multiple clients created in child process/worker; only the first will handle sharding helpers.',
+      );
     }
     return this._singleton;
   }
