@@ -1,95 +1,38 @@
-const Constants = require('../util/Constants');
-const Collection = require('../util/Collection');
+'use strict';
+
+const Base = require('./Base');
 const Snowflake = require('../util/Snowflake');
 
 /**
- * Represents a custom emoji.
+ * Represents an emoji, see {@link GuildEmoji} and {@link ReactionEmoji}.
+ * @extends {Base}
  */
-class Emoji {
-  constructor(guild, data) {
+class Emoji extends Base {
+  constructor(client, emoji) {
+    super(client);
     /**
-     * The client that instantiated this object
-     * @name Emoji#client
-     * @type {Client}
-     * @readonly
+     * Whether this emoji is animated
+     * @type {boolean}
      */
-    Object.defineProperty(this, 'client', { value: guild.client });
+    this.animated = emoji.animated;
 
     /**
-     * The guild this emoji is part of
-     * @type {Guild}
-     */
-    this.guild = guild;
-
-    this.setup(data);
-  }
-
-  setup(data) {
-    /**
-     * The ID of the emoji
-     * @type {Snowflake}
-     */
-    this.id = data.id;
-
-    /**
-     * The name of the emoji
+     * The name of this emoji
      * @type {string}
      */
-    this.name = data.name;
+    this.name = emoji.name;
 
     /**
-     * Whether or not this emoji requires colons surrounding it
-     * @type {boolean}
+     * The ID of this emoji
+     * @type {?Snowflake}
      */
-    this.requiresColons = data.require_colons;
+    this.id = emoji.id;
 
     /**
-     * Whether this emoji is managed by an external service
+     * Whether this emoji has been deleted
      * @type {boolean}
      */
-    this.managed = data.managed;
-
-    this._roles = data.roles;
-  }
-
-  /**
-   * The timestamp the emoji was created at
-   * @type {number}
-   * @readonly
-   */
-  get createdTimestamp() {
-    return Snowflake.deconstruct(this.id).timestamp;
-  }
-
-  /**
-   * The time the emoji was created
-   * @type {Date}
-   * @readonly
-   */
-  get createdAt() {
-    return new Date(this.createdTimestamp);
-  }
-
-  /**
-   * A collection of roles this emoji is active for (empty if all), mapped by role ID
-   * @type {Collection<Snowflake, Role>}
-   * @readonly
-   */
-  get roles() {
-    const roles = new Collection();
-    for (const role of this._roles) {
-      if (this.guild.roles.has(role)) roles.set(role, this.guild.roles.get(role));
-    }
-    return roles;
-  }
-
-  /**
-   * The URL to the emoji file
-   * @type {string}
-   * @readonly
-   */
-  get url() {
-    return Constants.Endpoints.CDN(this.client.options.http.cdn).Emoji(this.id);
+    this.deleted = false;
   }
 
   /**
@@ -98,117 +41,63 @@ class Emoji {
    * @readonly
    */
   get identifier() {
-    if (this.id) return `${this.name}:${this.id}`;
+    if (this.id) return `${this.animated ? 'a:' : ''}${this.name}:${this.id}`;
     return encodeURIComponent(this.name);
   }
 
   /**
-   * Data for editing an emoji.
-   * @typedef {Object} EmojiEditData
-   * @property {string} [name] The name of the emoji
-   * @property {Collection<Snowflake, Role>|Array<Snowflake|Role>} [roles] Roles to restrict emoji to
+   * The URL to the emoji file if its a custom emoji
+   * @type {?string}
+   * @readonly
    */
-
-  /**
-   * Edits the emoji.
-   * @param {EmojiEditData} data The new data for the emoji
-   * @param {string} [reason] Reason for editing this emoji
-   * @returns {Promise<Emoji>}
-   * @example
-   * // Edit an emoji
-   * emoji.edit({name: 'newemoji'})
-   *   .then(e => console.log(`Edited emoji ${e}`))
-   *   .catch(console.error);
-   */
-  edit(data, reason) {
-    return this.client.rest.methods.updateEmoji(this, data, reason);
+  get url() {
+    if (!this.id) return null;
+    return this.client.rest.cdn.Emoji(this.id, this.animated ? 'gif' : 'png');
   }
 
   /**
-   * Set the name of the emoji.
-   * @param {string} name The new name for the emoji
-   * @param {string} [reason] The reason for changing the emoji's name
-   * @returns {Promise<Emoji>}
+   * The timestamp the emoji was created at, or null if unicode
+   * @type {?number}
+   * @readonly
    */
-  setName(name, reason) {
-    return this.edit({ name }, reason);
+  get createdTimestamp() {
+    if (!this.id) return null;
+    return Snowflake.deconstruct(this.id).timestamp;
   }
 
   /**
-   * Add a role to the list of roles that can use this emoji.
-   * @param {Role} role The role to add
-   * @returns {Promise<Emoji>}
+   * The time the emoji was created at, or null if unicode
+   * @type {?Date}
+   * @readonly
    */
-  addRestrictedRole(role) {
-    return this.addRestrictedRoles([role]);
+  get createdAt() {
+    if (!this.id) return null;
+    return new Date(this.createdTimestamp);
   }
 
   /**
-   * Add multiple roles to the list of roles that can use this emoji.
-   * @param {Role[]} roles Roles to add
-   * @returns {Promise<Emoji>}
-   */
-  addRestrictedRoles(roles) {
-    const newRoles = new Collection(this.roles);
-    for (const role of roles) {
-      if (this.guild.roles.has(role.id)) newRoles.set(role.id, role);
-    }
-    return this.edit({ roles: newRoles });
-  }
-
-  /**
-   * Remove a role from the list of roles that can use this emoji.
-   * @param {Role} role The role to remove
-   * @returns {Promise<Emoji>}
-   */
-  removeRestrictedRole(role) {
-    return this.removeRestrictedRoles([role]);
-  }
-
-  /**
-   * Remove multiple roles from the list of roles that can use this emoji.
-   * @param {Role[]} roles Roles to remove
-   * @returns {Promise<Emoji>}
-   */
-  removeRestrictedRoles(roles) {
-    const newRoles = new Collection(this.roles);
-    for (const role of roles) {
-      if (newRoles.has(role.id)) newRoles.delete(role.id);
-    }
-    return this.edit({ roles: newRoles });
-  }
-
-  /**
-   * When concatenated with a string, this automatically returns the emoji mention rather than the object.
+   * When concatenated with a string, this automatically returns the text required to form a graphical emoji on Discord
+   * instead of the Emoji object.
    * @returns {string}
    * @example
-   * // Send an emoji:
-   * const emoji = guild.emojis.first();
+   * // Send a custom emoji from a guild:
+   * const emoji = guild.emojis.cache.first();
    * msg.reply(`Hello! ${emoji}`);
+   * @example
+   * // Send the emoji used in a reaction to the channel the reaction is part of
+   * reaction.message.channel.send(`The emoji used was: ${reaction.emoji}`);
    */
   toString() {
-    return this.requiresColons ? `<:${this.name}:${this.id}>` : this.name;
+    return this.id ? `<${this.animated ? 'a' : ''}:${this.name}:${this.id}>` : this.name;
   }
 
-  /**
-   * Whether this emoji is the same as another one.
-   * @param {Emoji|Object} other The emoji to compare it to
-   * @returns {boolean} Whether the emoji is equal to the given emoji or not
-   */
-  equals(other) {
-    if (other instanceof Emoji) {
-      return (
-        other.id === this.id &&
-        other.name === this.name &&
-        other.managed === this.managed &&
-        other.requiresColons === this.requiresColons
-      );
-    } else {
-      return (
-        other.id === this.id &&
-        other.name === this.name
-      );
-    }
+  toJSON() {
+    return super.toJSON({
+      guild: 'guildID',
+      createdTimestamp: true,
+      url: true,
+      identifier: true,
+    });
   }
 }
 
