@@ -1,12 +1,27 @@
-const Collection = require('../util/Collection');
-const Emoji = require('./Emoji');
+'use strict';
+
+const GuildEmoji = require('./GuildEmoji');
 const ReactionEmoji = require('./ReactionEmoji');
+const ReactionUserManager = require('../managers/ReactionUserManager');
+const Util = require('../util/Util');
 
 /**
  * Represents a reaction to a message.
  */
 class MessageReaction {
-  constructor(message, emoji, count, me) {
+  /**
+   * @param {Client} client The instantiating client
+   * @param {Object} data The data for the message reaction
+   * @param {Message} message The message the reaction refers to
+   */
+  constructor(client, data, message) {
+    /**
+     * The client that instantiated this message reaction
+     * @name MessageReaction#client
+     * @type {Client}
+     * @readonly
+     */
+    Object.defineProperty(this, 'client', { value: client });
     /**
      * The message that this reaction refers to
      * @type {Message}
@@ -17,35 +32,55 @@ class MessageReaction {
      * Whether the client has given this reaction
      * @type {boolean}
      */
-    this.me = me;
+    this.me = data.me;
 
     /**
-     * The number of people that have given the same reaction
-     * @type {number}
+     * A manager of the users that have given this reaction
+     * @type {ReactionUserManager}
      */
-    this.count = count || 0;
+    this.users = new ReactionUserManager(client, undefined, this);
 
-    /**
-     * The users that have given this reaction, mapped by their ID
-     * @type {Collection<Snowflake, User>}
-     */
-    this.users = new Collection();
+    this._emoji = new ReactionEmoji(this, data.emoji);
 
-    this._emoji = new ReactionEmoji(this, emoji.name, emoji.id);
+    this._patch(data);
+  }
+
+  _patch(data) {
+    // eslint-disable-next-line eqeqeq
+    if (this.count == undefined) {
+      /**
+       * The number of people that have given the same reaction
+       * @type {?number}
+       */
+      this.count = data.count;
+    }
   }
 
   /**
-   * The emoji of this reaction, either an Emoji object for known custom emojis, or a ReactionEmoji
+   * Removes all users from this reaction.
+   * @returns {Promise<MessageReaction>}
+   */
+  async remove() {
+    await this.client.api
+      .channels(this.message.channel.id)
+      .messages(this.message.id)
+      .reactions(this._emoji.identifier)
+      .delete();
+    return this;
+  }
+
+  /**
+   * The emoji of this reaction, either an GuildEmoji object for known custom emojis, or a ReactionEmoji
    * object which has fewer properties. Whatever the prototype of the emoji, it will still have
    * `name`, `id`, `identifier` and `toString()`
-   * @type {Emoji|ReactionEmoji}
+   * @type {GuildEmoji|ReactionEmoji}
    * @readonly
    */
   get emoji() {
-    if (this._emoji instanceof Emoji) return this._emoji;
+    if (this._emoji instanceof GuildEmoji) return this._emoji;
     // Check to see if the emoji has become known to the client
     if (this._emoji.id) {
-      const emojis = this.message.client.emojis;
+      const emojis = this.message.client.emojis.cache;
       if (emojis.has(this._emoji.id)) {
         const emoji = emojis.get(this._emoji.id);
         this._emoji = emoji;
@@ -56,37 +91,45 @@ class MessageReaction {
   }
 
   /**
-   * Removes a user from this reaction.
-   * @param {UserResolvable} [user=this.message.client.user] The user to remove the reaction of
-   * @returns {Promise<MessageReaction>}
+   * Whether or not this reaction is a partial
+   * @type {boolean}
+   * @readonly
    */
-  remove(user = this.message.client.user) {
-    const message = this.message;
-    const userID = this.message.client.resolver.resolveUserID(user);
-    if (!userID) return Promise.reject(new Error('Couldn\'t resolve the user ID to remove from the reaction.'));
-    return message.client.rest.methods.removeMessageReaction(
-      message, this.emoji.identifier, userID
-    );
+  get partial() {
+    return this.count === null;
   }
 
   /**
-   * Fetch all the users that gave this reaction. Resolves with a collection of users, mapped by their IDs.
-   * @param {number} [limit=100] The maximum amount of users to fetch, defaults to 100
-   * @returns {Promise<Collection<Snowflake, User>>}
+   * Fetch this reaction.
+   * @returns {Promise<MessageReaction>}
    */
-  fetchUsers(limit = 100) {
-    const message = this.message;
-    return message.client.rest.methods.getMessageReactionUsers(
-      message, this.emoji.identifier, limit
-    ).then(users => {
-      this.users = new Collection();
-      for (const rawUser of users) {
-        const user = this.message.client.dataManager.newUser(rawUser);
-        this.users.set(user.id, user);
-      }
-      this.count = this.users.size;
-      return this.users;
-    });
+  async fetch() {
+    const message = await this.message.fetch();
+    const existing = message.reactions.cache.get(this.emoji.id || this.emoji.name);
+    // The reaction won't get set when it has been completely removed
+    this._patch(existing || { count: 0 });
+    return this;
+  }
+
+  toJSON() {
+    return Util.flatten(this, { emoji: 'emojiID', message: 'messageID' });
+  }
+
+  _add(user) {
+    if (this.partial) return;
+    this.users.cache.set(user.id, user);
+    if (!this.me || user.id !== this.message.client.user.id || this.count === 0) this.count++;
+    if (!this.me) this.me = user.id === this.message.client.user.id;
+  }
+
+  _remove(user) {
+    if (this.partial) return;
+    this.users.cache.delete(user.id);
+    if (!this.me || user.id !== this.message.client.user.id) this.count--;
+    if (user.id === this.message.client.user.id) this.me = false;
+    if (this.count <= 0 && this.users.cache.size === 0) {
+      this.message.reactions.cache.delete(this.emoji.id || this.emoji.name);
+    }
   }
 }
 
