@@ -1,31 +1,23 @@
+'use strict';
+
 const GuildChannel = require('./GuildChannel');
+const { Error } = require('../errors');
 const Collection = require('../util/Collection');
+const { browser } = require('../util/Constants');
+const Permissions = require('../util/Permissions');
 
 /**
  * Represents a guild voice channel on Discord.
  * @extends {GuildChannel}
  */
 class VoiceChannel extends GuildChannel {
-  constructor(guild, data) {
-    super(guild, data);
-
-    /**
-     * The members in this voice channel
-     * @type {Collection<Snowflake, GuildMember>}
-     */
-    this.members = new Collection();
-
-    this.type = 'voice';
-  }
-
-  setup(data) {
-    super.setup(data);
-
+  _patch(data) {
+    super._patch(data);
     /**
      * The bitrate of this voice channel
      * @type {number}
      */
-    this.bitrate = data.bitrate * 0.001;
+    this.bitrate = data.bitrate;
 
     /**
      * The maximum amount of users allowed in this channel - 0 means unlimited.
@@ -35,14 +27,18 @@ class VoiceChannel extends GuildChannel {
   }
 
   /**
-   * The voice connection for this voice channel, if the client is connected
-   * @type {?VoiceConnection}
+   * The members in this voice channel
+   * @type {Collection<Snowflake, GuildMember>}
    * @readonly
    */
-  get connection() {
-    const connection = this.guild.voiceConnection;
-    if (connection && connection.channel.id === this.id) return connection;
-    return null;
+  get members() {
+    const coll = new Collection();
+    for (const state of this.guild.voiceStates.cache.values()) {
+      if (state.channelID === this.id && state.member) {
+        coll.set(state.id, state.member);
+      }
+    }
+    return coll;
   }
 
   /**
@@ -55,14 +51,33 @@ class VoiceChannel extends GuildChannel {
   }
 
   /**
-   * Checks if the client has permission join the voice channel
+   * Whether the channel is deletable by the client user
+   * @type {boolean}
+   * @readonly
+   */
+  get deletable() {
+    return super.deletable && this.permissionsFor(this.client.user).has(Permissions.FLAGS.CONNECT, false);
+  }
+
+  /**
+   * Whether the channel is editable by the client user
+   * @type {boolean}
+   * @readonly
+   */
+  get editable() {
+    return this.manageable && this.permissionsFor(this.client.user).has(Permissions.FLAGS.CONNECT, false);
+  }
+
+  /**
+   * Whether the channel is joinable by the client user
    * @type {boolean}
    * @readonly
    */
   get joinable() {
-    if (this.client.browser) return false;
-    if (!this.permissionsFor(this.client.user).has('CONNECT')) return false;
-    if (this.full && !this.permissionsFor(this.client.user).has('MOVE_MEMBERS')) return false;
+    if (browser) return false;
+    if (!this.viewable) return false;
+    if (!this.permissionsFor(this.client.user).has(Permissions.FLAGS.CONNECT, false)) return false;
+    if (this.full && !this.permissionsFor(this.client.user).has(Permissions.FLAGS.MOVE_MEMBERS, false)) return false;
     return true;
   }
 
@@ -72,22 +87,21 @@ class VoiceChannel extends GuildChannel {
    * @readonly
    */
   get speakable() {
-    return this.permissionsFor(this.client.user).has('SPEAK');
+    return this.permissionsFor(this.client.user).has(Permissions.FLAGS.SPEAK, false);
   }
 
   /**
-   * Sets the bitrate of the channel (in kbps).
+   * Sets the bitrate of the channel.
    * @param {number} bitrate The new bitrate
    * @param {string} [reason] Reason for changing the channel's bitrate
    * @returns {Promise<VoiceChannel>}
    * @example
    * // Set the bitrate of a voice channel
-   * voiceChannel.setBitrate(48)
-   *   .then(vc => console.log(`Set bitrate to ${vc.bitrate}kbps for ${vc.name}`))
+   * voiceChannel.setBitrate(48000)
+   *   .then(vc => console.log(`Set bitrate to ${vc.bitrate}bps for ${vc.name}`))
    *   .catch(console.error);
    */
   setBitrate(bitrate, reason) {
-    bitrate *= 1000;
     return this.edit({ bitrate }, reason);
   }
 
@@ -116,7 +130,7 @@ class VoiceChannel extends GuildChannel {
    *   .catch(console.error);
    */
   join() {
-    if (this.client.browser) return Promise.reject(new Error('Voice connections are not available in browsers.'));
+    if (browser) return Promise.reject(new Error('VOICE_NO_BROWSER'));
     return this.client.voice.joinChannel(this);
   }
 
@@ -127,7 +141,7 @@ class VoiceChannel extends GuildChannel {
    * voiceChannel.leave();
    */
   leave() {
-    if (this.client.browser) return;
+    if (browser) return;
     const connection = this.client.voice.connections.get(this.guild.id);
     if (connection && connection.channel.id === this.id) connection.disconnect();
   }
