@@ -1,3 +1,7 @@
+'use strict';
+
+const Util = require('../util/Util');
+
 /**
  * Represents a Discord voice region for guilds.
  */
@@ -38,12 +42,10 @@ class VoiceRegion {
      * @type {boolean}
      */
     this.custom = data.custom;
+  }
 
-    /**
-     * A sample hostname for what a connection might look like
-     * @type {string}
-     */
-    this.sampleHostname = data.sample_hostname;
+  toJSON() {
+    return Util.flatten(this);
   }
 }
 
